# docs/api/main-backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RoomPACT Main Backend",
    "description": "Main backend API for RoomPACT Campus",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Public API behind nginx on EC2"
    },
    {
      "url": "http://localhost:8000",
      "description": "Direct local development"
    }
  ],
  "paths": {
...
    "/profiles": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create Profile",
        "operationId": "create_profile_profiles_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Profile Profiles Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Profiles Profiles Get"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/{profile_id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Profile Profiles  Profile Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_profiles__profile_id__put",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Profile Profiles  Profile Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/{profile_id}/interview": {
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Save Profile Interview",
        "operationId": "save_profile_interview_profiles__profile_id__interview_put",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInterviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Profile Interview Profiles  Profile Id  Interview Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get Profile Interview",
        "operationId": "get_profile_interview_profiles__profile_id__interview_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Profile Interview Profiles  Profile Id  Interview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/{profile_id}/recommendations": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get Profile Recommendations",
        "operationId": "get_profile_recommendations_profiles__profile_id__recommendations_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Profile Recommendations Profiles  Profile Id  Recommendations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileCreateRequest": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Nickname"
          },
          "age": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 17.0,
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Gender"
          },
          "region": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Region"
          },
          "move_in_period": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Move In Period"
          },
          "stay_duration_months": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Stay Duration Months"
          }
        },
        "type": "object",
        "required": [
          "nickname",
          "age",
          "gender",
          "region",
          "move_in_period",
          "stay_duration_months"
        ],
        "title": "ProfileCreateRequest"
      },
      "ProfileInterviewRequest": {
        "properties": {
          "wake_up_time": {
            "type": "string",
            "pattern": "^(?:[01]\\d|2[0-3]):(?:00|10|20|30|40|50)$",
            "title": "Wake Up Time"
          },
          "sleep_time": {
            "type": "string",
            "pattern": "^(?:[01]\\d|2[0-3]):(?:00|10|20|30|40|50)$",
            "title": "Sleep Time"
          },
          "noise_sensitive": {
            "type": "boolean",
            "title": "Noise Sensitive"
          },
          "quiet_hours_start": {
            "type": "string",
            "pattern": "^(?:[01]\\d|2[0-3]):(?:00|10|20|30|40|50)$",
            "title": "Quiet Hours Start"
          },
          "cleaning_frequency": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "매일"
            ],
            "title": "Cleaning Frequency"
          },
          "dishes_deadline": {
            "type": "string",
            "enum": [
              "바로",
              "그날 이내에",
              "다음날 아침"
            ],
            "title": "Dishes Deadline"
          },
          "guest_frequency": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "매일"
            ],
            "title": "Guest Frequency"
          },
          "smokes": {
            "type": "boolean",
            "title": "Smokes"
          },
          "smoking_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Smoking Type"
          },
          "smoking_place": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "밖",
                  "베란다",
                  "집 안"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Smoking Place"
          },
          "drinking_frequency": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "매일"
            ],
            "title": "Drinking Frequency"
          },
          "home_stay_frequency": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "매일"
            ],
            "title": "Home Stay Frequency"
          },
          "meal_preference": {
            "type": "string",
            "enum": [
              "배달",
              "직접"
            ],
            "title": "Meal Preference"
          },
          "home_activity_frequency": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "매일"
            ],
            "title": "Home Activity Frequency"
          },
          "supplies_sharing": {
            "type": "string",
            "enum": [
              "공동구매",
              "각자",
              "일부 공유"
            ],
            "title": "Supplies Sharing"
          },
          "summer_temperature": {
            "type": "integer",
            "maximum": 35.0,
            "minimum": 16.0,
            "title": "Summer Temperature"
          },
          "winter_temperature": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 10.0,
            "title": "Winter Temperature"
          },
          "pet_ok": {
            "type": "boolean",
            "title": "Pet Ok"
          },
          "pet_preference": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "고양이",
                  "강아지",
                  "둘 다"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Pet Preference"
          },
          "conflict_resolution": {
            "type": "string",
            "enum": [
              "즉시 대면",
              "모아서 대면"
            ],
            "title": "Conflict Resolution"
          },
          "shared_cost_rule": {
            "type": "string",
            "enum": [
              "반반",
              "거주 시간 비율"
            ],
            "title": "Shared Cost Rule"
          },
          "personal_space_access": {
            "type": "string",
            "enum": [
              "자유롭게",
              "노크 혹은 허락",
              "불가능"
            ],
            "title": "Personal Space Access"
          },
          "personal_space_ratio": {
            "type": "string",
            "enum": [
              "반반",
              "필요한 만큼"
            ],
            "title": "Personal Space Ratio"
          },
          "security_preference": {
            "type": "string",
            "enum": [
              "항시 잠금",
              "외출시",
              "상관없음"
            ],
            "title": "Security Preference"
          },
          "absence_notice": {
            "type": "string",
            "enum": [
              "항상",
              "하루 이상",
              "필요 없음"
            ],
            "title": "Absence Notice"
          },
          "hardcut_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Hardcut Conditions"
          }
        },
        "type": "object",
        "required": [
          "wake_up_time",
          "sleep_time",
          "noise_sensitive",
          "quiet_hours_start",
          "cleaning_frequency",
          "dishes_deadline",
          "guest_frequency",
          "smokes",
          "drinking_frequency",
          "home_stay_frequency",
          "meal_preference",
          "home_activity_frequency",
          "supplies_sharing",
          "summer_temperature",
          "winter_temperature",
          "pet_ok",
          "conflict_resolution",
          "shared_cost_rule",
          "personal_space_access",
          "personal_space_ratio",
          "security_preference",
          "absence_notice"
        ],
        "title": "ProfileInterviewRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}